+++ src/app/app.component.html
@@ -1,3 +1,3 @@
 <app-header/>
-<app-form (data)="calculateInvestmentResults($event)"/>
-<app-result [result]="resultData()"/>
+<app-form/>
+<app-result/>
