+++ src/app/component/header/header.component.html
@@ -2,6 +2,6 @@
     <img src="/public/codvary-logo.png" alt="investment calculator"/>
     <div>
         <h1>Investment Calculator</h1>
-        <p>your feture forcasting on demand</p>
+        <p>your future forcasting on demand</p>
     </div>
 </header>
